reconnected moved edit and delete buttons to event listeners

+++ public/javascript/dash-delete.js
@@ -1,19 +1,25 @@
 async function deletePostDashboard(event) {
     event.preventDefault();
 
-    const post_id = document.querySelector('.btn-delete-dashboard').getAttribute('data-id');
-    console.log(post_id);
+    const post_id = event.target.getAttribute('data-id');
+    //console.log('data-id'+ post_id);
 
-    const response = await fetch(`/api/posts/${post_id}`, {
+    if(post_id) {
+        const response = await fetch(`/api/posts/${post_id}`, {
         method: 'DELETE'
-    });
+        });
 
-    if (response.ok) {
+        if (response.ok) {
         document.location.replace('/dashboard');
+        }
+        else {
+            alert(response.statusText);
+        }
     }
     else {
-        alert(response.statusText);
+        const edit_post_id = event.target.getAttribute('data-post-id');
+        window.location.replace(`/dashboard/edit/${edit_post_id}`);
     }
 }
 
-document.querySelector('.btn-delete-dashboard').addEventListener('click', deletePostDashboard);
+document.querySelectorAll('.btn-edit-container').forEach(btn => btn.addEventListener('click', deletePostDashboard));
